feat: Add tsparticles package and implement particle background
This commit adds the "@tsparticles/engine" package to the project dependencies and implements a particle background using the tsparticles library. The "Particles.astro" component is created to handle the rendering of the particles, and the "particles.json" file is added to the "src/assets" directory to provide the particle configuration. The particle background enhances the visual appeal of the website.